Move routes testing code into a file for each specific route
Hoist!

Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
