+++ tests/main.rs
@@ -1,0 +1,7 @@
+#![feature(proc_macro_hygiene, decl_macro)]
+
+#[macro_use]
+extern crate rocket;
+extern crate data_server;
+
+mod routes;
