+++ tests/routes/mod.rs
@@ -1,0 +1,8 @@
+// use super::*;
+
+use rocket::http::Status;
+use rocket::local::Client;
+
+use data_server::routes::*;
+
+mod ping;
